requirements.txt
livekit-agents>=1.2.1
livekit-plugins-google>=1.0.0
livekit-plugins-noise-cancellation>=0.2.0
python-dotenv